pyproject.toml
[tool.poetry]
name = "avabot-backend"
version = "0.1.0"
description = ""
authors = ["[chideracollins]' <[siralexcollins@gmail.com]>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
langchain = "^0.3.7"
langchain-google-genai = "^2.0.4"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
flask-cors = "^5.0.0"
httpx = "^0.28.1"
pillow = "^11.1.0"


[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true
without = ["dev"]


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
